# Avoid copy of std::function-based callback in path unpacking